8195100: Use a low latency hashtable for SymbolTable
Used concurrentHashTable, similar to stringTable

+++ src/hotspot/share/oops/symbol.cpp
@@ -396,4 +396,4 @@ bool Symbol::is_valid(Symbol* s) {
 }
 
 // SymbolTable prints this in its statistics
-NOT_PRODUCT(int Symbol::_total_count = 0;)
+NOT_PRODUCT(size_t Symbol::_total_count = 0;)
